feat: Implement feedback retrieval and validation enhancements, including customer-specific access and response DTO

+++ Domain/ReadModels/FeedbackReadModel.cs
@@ -4,7 +4,7 @@ public class FeedbackReadModel
 {
     public Ulid FeedbackId { get; set; }
     public string? Substance { get; set; }
-    public double RatingScore { get; set; }
+    public float RatingScore { get; set; }
     public string? ImageUrl { get; set; }
     public Ulid BillId { get; set; }
     public Ulid CustomerId { get; set; }
